% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{barplot_3d}
\alias{barplot_3d}
\title{Generate 3d barplot.}
\usage{
barplot_3d(
  x,
  y,
  z,
  col = NULL,
  zvec = c(0.1, 0.1),
  alpha = 0.8,
  border = "grey",
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{x}{a numeric vector of proportions that will be shown on the x axis}

\item{y}{a numeric vector of heights that will be shown on the y axis}

\item{z}{a numeric vector of depths}

\item{col}{a numeric vector of colors}

\item{zvec}{a numeric vector of length 2 for the direction of depth. Both entries should be positive.}

\item{alpha}{a transparency value taking value between 0 and 1, where 1 is opaque and 0 is transparent}

\item{border}{the border color}

\item{xlim}{user-specified xlim}

\item{ylim}{user-specified ylim}
}
\description{
This function is used to illustrate the model-based mean expression in the CARseq paper.
}
\examples{
barplot_3d(x = 1:3, y = 1:3, z = 1)

}
