% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\docType{package}
\name{CARseq-package}
\alias{CARseq-package}
\title{CARseq package for cell type-specific differential analysis of count data}
\description{
CARseq package for cell type-specific differential analysis of count data
}
\references{

}
\author{
Chong Jin, Wei Sun
}
\keyword{package}
