% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_shrinkage_prior.R
\name{estimate_shrinkage_prior}
\alias{estimate_shrinkage_prior}
\title{Estimate the prior needed for LFC shrinkage}
\usage{
estimate_shrinkage_prior(estimates, K, M, H, prob = 0.95)
}
\arguments{
\item{estimates}{A matrix of K (number of cell type-independent variables) + H (number of cell types) * M (number of groups) + 1 (overdispersion)
columns. The estimates need to be obtained without using LFC shrinkage.}

\item{K}{number of cell type-independent variables}

\item{M}{number of group}

\item{H}{number of cell types}

\item{prob}{a number to match \code{prob} quantile of absolute value of 
parameter estimates to \code{prob} confidence interval}
}
\value{
a vector of length \eqn{K + H * (M + 1)}. The first K entries are 0,
the next H * M entries are based on 1/sigma^2 of length H of the normal prior 
for each cell type-specific LFC, and the last M entries are based on
1/sigma^2 of length 1 of the normal prior for
the cell type-specific intercept terms.
}
\description{
\code{estimate_shrinkage_prior}, for each cell type, matches the 95 percentile of absolute values of
all log fold changes (LFCs)
to the 97.5 percentile of a zero-centered normal distribution. It also matches the 95 percentile of 
absolute values of all cell type-specific expressions on log scale
to the 97.5 percentile of a zero-centered normal distribution.
}
