% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_shrinkage_prior.R
\name{estimate_variance_parameter}
\alias{estimate_variance_parameter}
\title{Estimate variance parameter of zero-mean normal prior based on a vector of numbers}
\usage{
estimate_variance_parameter(estimates, prob = 0.95)
}
\arguments{
\item{estimates}{a vector (or a matrix) of numbers from which to estimate normal prior parameters}

\item{prob}{a number to match \code{prob} quantile of \code{abs(x)} to \code{prob} confidence interval
of normal distribution}
}
\value{
variance parameter sigma^2 (or variance parameters for each column of a matrix)
of zero-mean normal distribution
}
\description{
Estimate variance parameter of zero-mean normal prior based on a vector of numbers
}
