% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_qvalues_one_inflated}
\alias{get_qvalues_one_inflated}
\title{Calculate q-values (Storey 2003) from a p-value distribution with considerable amount of 1's}
\usage{
get_qvalues_one_inflated(pvals)
}
\arguments{
\item{pvals}{a numeric vector of p-values.}
}
\description{
Calculate q-values (Storey 2003) from a p-value distribution with considerable amount of 1's
}
\examples{
set.seed(1234)
pval = c(runif(100, 0, 0.1), runif(100,0, 1), rep(1, 100))
qval = get_qvalues_one_inflated(pval)
plot(sort(pval), sort(qval), xlim=c(0,1), ylim=c(0,1), type="l",
     xlab="p-values", ylab="q-values")
qval_range = seq(0.001, 0.25, by=0.001)
plot(qval_range,
     sapply(qval_range, function(x) sum(qval <= x, na.rm=TRUE)),
     type="l",
     xlab="q-value", ylab="number of significant genes")
}
