% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permute_case_and_controls}
\alias{permute_case_and_controls}
\title{Permute case and control labels symmetrically}
\usage{
permute_case_and_controls(labels)
}
\arguments{
\item{labels}{a character vector of labels with exactly two levels.}
}
\description{
Permute case and control labels symmetrically
}
