% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{run_CARseq}
\alias{run_CARseq}
\title{Conduct a CARseq test}
\usage{
run_CARseq(
  count_matrix,
  cellular_proportions,
  groups,
  formula = NULL,
  data = NULL,
  read_depth = 1,
  shrinked_lfc = TRUE,
  cores = 1
)
}
\arguments{
\item{count_matrix}{A matrix of G x n total read counts observed.}

\item{cellular_proportions}{A matrix of n x H of cellular proportions.}

\item{groups}{A vector of length n indicating groups that we would like to test. Will be coerced to be factors.}

\item{formula}{A formula of an intercept term "1" and other cell type-indepedent variables. Can be NULL.}

\item{data}{A data frame containing the cell type-indepedent variables specified in \code{formula}. Can be NULL.}

\item{read_depth}{A vector of n sample-specific read depths. It it used as an offset term in the
regression model. Alternatively, it can be 1, NA or NULL so that we don't have an offset term
in the model, and log(read_depth) can be included as one of the cell type-independent variables.}

\item{shrinked_lfc}{Logical. If \code{TRUE} (default), provide shrinked log fold change for cell type-specific variables.}

\item{cores}{Numeric. Number of cores to use in \code{doMC::registerDoMC} which uses system \code{fork} call.}
}
\value{
Returns a list mostly of matrices. Note that the matrices with "shrinked" in their names are only available when \code{shrinked_lfc} is \code{TRUE}: 
\describe{
  \item{p}{A matrix of G x H p-values}
  \item{padj}{A matrix of G x H p-values adjusted using Benjamini & Hochberg (1995).}
  \item{shrinked_lfc}{A matrix of shrinked log fold change between cell type-specific effects.}
  \item{shrinked_lfcSE}{A matrix of standard errors of shrinked log fold change of cell type-specific effects between different groups.}
  \item{shrinked_coefficients}{A matrix of shrinked coefficient estimates.}
  \item{shrinked_coefficientsSE}{A matrix of standard errors of shrinked coefficient estimates.}
  \item{lfc}{A matrix of log fold change between cell type-specific effects.}
  \item{lfcSE}{A matrix of standard errors of log fold change of cell type-specific effects between between different groups of cell type-specific effects.}
  \item{coefficients}{A matrix of MLE coefficient estimates.}
  \item{coefficientsSE}{A matrix of standard errors of coefficient estimates.}
  \item{overdispersion}{A matrix of overdispersion parameters. The overdispersion parameter
                        is parametrized so that the variance of the negative binomial distribution is \code{mean + mean^2/overdispersion}.}
  \item{elapsed_time}{The time elapsed.}
}
}
\description{
Conduct a CARseq test
}
\examples{
data("n50DE221rep1")
# As an example, run only the first 10 genes among the all 10,000 genes to save time:
res = run_CARseq(count_matrix = n50DE221rep1$observed_read_count[1:10,],
                 cellular_proportions = n50DE221rep1$rho,
                 groups = gl(2, 25),
                 formula = ~ RIN,
                 data = n50DE221rep1$clinical_variables,
                 read_depth = n50DE221rep1$d,
                 shrinked_lfc = TRUE,
                 cores = 2
)

}
