% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{run_CARseq}
\alias{run_CARseq}
\title{Conduct a CARseq test}
\usage{
run_CARseq(
  count_matrix,
  cellular_proportions,
  groups,
  formula = NULL,
  data = NULL,
  read_depth = 1,
  shrunken_lfc = TRUE,
  cores = 1,
  fix_overdispersion = FALSE,
  useSocket = TRUE
)
}
\arguments{
\item{count_matrix}{A matrix of G x n total read counts observed.}

\item{cellular_proportions}{A matrix of n x H of cellular proportions.}

\item{groups}{A vector of length n indicating groups that we would like to test. Will be coerced to be factors.}

\item{formula}{A formula of an intercept term "1" and other cell type-independent variables. Can be NULL.}

\item{data}{A data frame containing the cell type-indepedent variables specified in \code{formula}. Can be NULL.}

\item{read_depth}{A vector of n sample-specific read depths. It it used as an offset term in the
regression model. Alternatively, it can be 1, NA or NULL so that we don't have an offset term
in the model, and log(read_depth) can be included as one of the cell type-independent variables.}

\item{shrunken_lfc}{Logical. If \code{TRUE} (default), provide shrunken log fold change for cell type-specific variables.}

\item{cores}{Numeric. Number of cores to use in \code{parallel::makePSOCKcluster}.
Note that for faster execution of the package, OPENBLAS or MKL library is recommended.
When OPENBLAS is used, add environment variable \code{Sys.setenv(OPENBLAS_NUM_THREADS=1)}.
When MKL is used,  add environment variables \code{Sys.setenv(MKL_NUM_THREADS=1)} and \code{Sys.setenv(MKL_THREADING_LAYER="GNU")}.}

\item{fix_overdispersion}{Logical or numeric. In general, when the sample size is sufficiently large (for example 10 samples per degree of freedom),
fix_overdispersion should be FALSE so that the overdispersion parameter is re-estimated in the reduced model.
However, when sample size is smaller, overdispersion parameter is hard to estimate and it is might be advantageous to
always use overdispersion parameter estimated under the full model, which is similar to how DESeq2 LRT works.
Alternatively, only for experimental purposes, 
a list of overdispersion parameters equal to the number of samples,
parametrized so that the variance of the negative binomial distribution is \code{mean + mean^2/overdispersion},
can be provided so that pre-computed overdispersion parameters can be used.}

\item{useSocket}{If \code{TRUE} (default), use socket for parallel computation, which works on Windows
as well as on POSIX systems (Mac, Linux, Unix, BSD) most of the time.
If \code{FALSE}, use fork for parallele computation, which on POSIX systems
and not Windows.}
}
\value{
Returns a list mostly of matrices. Note that the matrices with "shrunken" in their names are only available when \code{shrunken_lfc} is \code{TRUE}: 
\describe{
  \item{p}{A matrix of G x H p-values}
  \item{padj}{A matrix of G x H p-values adjusted using Benjamini & Hochberg (1995).}
  \item{shrunken_lfc}{A matrix of shrunken log fold change between cell type-specific effects.}
  \item{shrunken_lfcSE}{A matrix of standard errors of shrunken log fold change of cell type-specific effects between different groups.}
  \item{shrunken_coefficients}{A matrix of shrunken coefficient estimates.}
  \item{shrunken_coefficientsSE}{A matrix of standard errors of shrunken coefficient estimates.}
  \item{lfc}{A matrix of log fold change between cell type-specific effects.}
  \item{lfcSE}{A matrix of standard errors of log fold change of cell type-specific effects between between different groups of cell type-specific effects.}
  \item{coefficients}{A matrix of MLE coefficient estimates.}
  \item{coefficientsSE}{A matrix of standard errors of coefficient estimates.}
  \item{overdispersion}{A matrix of overdispersion parameters. The overdispersion parameter
                        is parametrized so that the variance of the negative binomial distribution is \code{mean + mean^2/overdispersion}.}
  \item{elapsed_time}{The time elapsed.}
}
}
\description{
Conduct a CARseq test
}
\examples{
data("n50DE221rep1")
# As an example, run only the first 10 genes among the all 10,000 genes to save time:
res = run_CARseq(count_matrix = n50DE221rep1$observed_read_count[1:10,],
                 cellular_proportions = n50DE221rep1$rho,
                 groups = gl(2, 25),
                 formula = ~ RIN,
                 data = n50DE221rep1$clinical_variables,
                 read_depth = n50DE221rep1$d,
                 shrunken_lfc = TRUE,
                 cores = 1,
                 fix_overdispersion = FALSE,
                 useSocket = TRUE
)

}
